speak-to-llm/requirements.txt
# Core dependencies for speech processing
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.4.0
pyaudio>=0.2.11

# LLM and AI libraries
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0
langchain>=0.0.350
langchain-openai>=0.0.2

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0

# Web framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: ElevenLabs for premium TTS
elevenlabs>=0.2.0

# Optional: Gradio for web interface
gradio>=4.0.0